refactor(parser): slim stmt/mod.rs by extracting submodules (§7-8)
Reduce the 2398-line parser/stmt/mod.rs (over the 500-line limit) to a thin
243-line root by moving the large #[cfg(test)] mod tests block and cohesive
groups of free functions into new sibling submodules, all under 500 lines.
mod.rs keeps its existing mod/use/re-export declarations, STMT_PARSERS,
memo/preseed setters, and the core statement/program entry points; new
submodules are re-exported preserving original visibility, so no external
caller changes.

- mod.rs (243)       - root: entry points, memo/preseed, mod/use/re-exports
- stmtlist.rs (356)  - block / statement-list parsing
- idents.rs (261)    - keyword / identifier / var-name parsing
- pub_shims.rs (116) - the thin *_pub delegation wrappers
- tests*.rs (376/363/364/388) - former inline #[cfg(test)] mod, split by range

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; one path qualified to absolute. All 121 unit
tests preserved (131 pass). The fn multiset is byte-identical. No behavior change.
