+++ src/parser/stmt/mod.rs
@@ -3,9 +3,12 @@ pub(crate) mod assign;
 pub(super) mod class;
 pub(in crate::parser) mod control;
 pub(crate) mod decl;
+mod idents;
 pub(super) mod modifier;
+mod pub_shims;
 pub(super) mod simple;
 mod simple_expr_stmt;
+mod stmtlist;
 mod sub;
 pub(crate) mod sub_param;
 
@@ -29,6 +32,28 @@ use class::class_decl_body;
 use modifier::{is_stmt_modifier_keyword, parse_statement_modifier};
 use sub::{
     method_decl_body, method_decl_body_my, parse_param_list, parse_sub_traits, sub_decl_body,
+};
+
+// Re-export the identifier/keyword/variable-name parsers moved to `idents`.
+// `keyword`/`parse_raku_ident` are referenced from sibling parser modules
+// (`primary`, `expr`) and must keep their original `pub(super)` reach; the
+// others are only used within the `stmt` descendant tree.
+use idents::{ident, qualified_ident, starts_unit_class_role_grammar, var_name};
+pub(super) use idents::{keyword, parse_raku_ident};
+
+// Re-export the block/statement-list parsers moved to `stmtlist`.
+use stmtlist::block;
+pub(in crate::parser) use stmtlist::stmt_list_pub;
+pub(super) use stmtlist::{block_inner, stmt_list_partial};
+
+// Re-export the thin public-accessor shims moved to `pub_shims`, all at their
+// original `pub(super)` visibility (consumed by `primary`/`expr`).
+pub(super) use pub_shims::{
+    constant_decl_pub, for_stmt_pub, ident_pub, if_stmt_pub, labeled_loop_stmt_pub,
+    lazy_for_stmt_pub, let_stmt_pub, loop_stmt_pub, my_decl_expr_pub, parse_param_list_pub,
+    parse_param_list_with_return_pub, parse_pointy_param_pub, parse_return_type_annotation_pub,
+    parse_sub_name_pub, parse_sub_traits_pub, statement_pub, sub_decl_with_semicolon_mode_pub,
+    temp_stmt_pub, until_stmt_pub, while_stmt_pub,
 };
 
 thread_local! {
@@ -64,713 +89,6 @@ pub(super) fn set_eval_user_sub_preseed(names: Vec<String>) {
 
 pub(super) fn statement_memo_stats() -> (usize, usize, usize) {
     STMT_MEMO.stats()
-}
-
-/// Try to match a keyword at the start of input, ensuring word boundary.
-pub(super) fn keyword<'a>(kw: &str, input: &'a str) -> Option<&'a str> {
-    if input.starts_with(kw) && !is_ident_char(input.as_bytes().get(kw.len()).copied()) {
-        Some(&input[kw.len()..])
-    } else {
-        None
-    }
-}
-
-/// True when `input` begins a `unit class`/`unit role`/`unit grammar`
-/// declaration (the semicolon, rest-of-unit form), whose following statements
-/// must be captured as the type's body. `unit module`/`unit package` are
-/// deliberately excluded — their package context is set compiler-side instead.
-fn starts_unit_class_role_grammar(input: &str) -> bool {
-    let Some(r) = keyword("unit", input) else {
-        return false;
-    };
-    let Ok((r, _)) = ws1(r) else {
-        return false;
-    };
-    keyword("class", r).is_some() || keyword("role", r).is_some() || keyword("grammar", r).is_some()
-}
-
-/// Parse an identifier (alphanumeric, _, -).
-/// Hyphen is only allowed when followed by an alphabetic char (not a digit).
-fn ident(input: &str) -> PResult<'_, String> {
-    let (rest, name) = parse_raku_ident(input)?;
-    Ok((rest, normalize_raku_identifier(name)))
-}
-
-/// Parse a Raku-style identifier.
-/// Allows hyphens and apostrophes between word segments, but only when followed by a letter.
-/// e.g. `foo-bar` is valid, `doesn't` is valid, but `foo-3` stops at `foo`.
-pub(super) fn parse_raku_ident<'a>(input: &'a str) -> PResult<'a, &'a str> {
-    // Check first character: underscore or Unicode alphabetic character.
-    let first = input
-        .chars()
-        .next()
-        .ok_or_else(|| PError::expected("identifier"))?;
-    if !is_raku_identifier_start(first) {
-        return Err(PError::expected("identifier"));
-    }
-    let mut end = first.len_utf8();
-    // Continue consuming identifier characters
-    let mut chars = input[end..].chars().peekable();
-    while let Some(&c) = chars.peek() {
-        if is_raku_identifier_continue(c) {
-            end += c.len_utf8();
-            chars.next();
-        } else if c == '-' || c == '\'' {
-            // Hyphen/apostrophe is part of identifier only if followed by a letter
-            let mut lookahead = chars.clone();
-            lookahead.next(); // skip the hyphen/apostrophe
-            if let Some(&next) = lookahead.peek() {
-                if is_raku_identifier_start(next) {
-                    end += 1; // consume the hyphen/apostrophe
-                    chars.next();
-                } else {
-                    break;
-                }
-            } else {
-                break;
-            }
-        } else {
-            break;
-        }
-    }
-    Ok((&input[end..], &input[..end]))
-}
-
-/// Parse a qualified identifier (Foo::Bar::Baz).
-/// `X::Syntax::Name::Null` — a qualified name with an empty component, e.g.
-/// `Foo::::Bar` or `$a::::b`.
-fn name_null_error() -> PError {
-    let msg = "Name component may not be null".to_string();
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), crate::value::Value::str(msg.clone()));
-    let ex = crate::value::Value::make_instance(
-        crate::symbol::Symbol::intern("X::Syntax::Name::Null"),
-        attrs,
-    );
-    PError::fatal_with_exception(msg, Box::new(ex))
-}
-
-fn qualified_ident(input: &str) -> PResult<'_, String> {
-    let (mut rest, name) = ident(input)?;
-    let mut full = name;
-    while rest.starts_with("::") {
-        let r = &rest[2..];
-        // `Foo::::Bar` / `$a::::b` — an empty component between `::` separators
-        // is a null name component, which Raku rejects at compile time.
-        if r.starts_with("::") {
-            return Err(name_null_error());
-        }
-        // Handle ::<SYMBOL> subscript syntax (e.g., CORE::<&run>)
-        if let Some(after_bracket) = r.strip_prefix('<')
-            && let Some(end) = after_bracket.find('>')
-        {
-            let symbol = &after_bracket[..end];
-            full.push_str("::");
-            full.push_str(symbol);
-            rest = &after_bracket[end + 1..];
-            continue;
-        }
-        if let Ok((r2, part)) = ident(r) {
-            full.push_str("::");
-            full.push_str(&part);
-            rest = r2;
-        } else {
-            return Err(PError::expected_at("identifier after '::'", r));
-        }
-    }
-    Ok((rest, full))
-}
-
-/// Parse a variable name ($x, @arr, %hash) and return just the name part.
-fn var_name(input: &str) -> PResult<'_, String> {
-    if input.starts_with('$')
-        || input.starts_with('@')
-        || input.starts_with('%')
-        || input.starts_with('&')
-    {
-        let r = &input[1..];
-        // Detect Perl 5 special variables before parsing twigils.
-        if input.starts_with('$')
-            && let Some(err) = super::primary::var::detect_perl5_scalar_var(r)
-        {
-            return Err(PError::fatal(err));
-        }
-        // Handle twigils
-        let (r, twigil) =
-            if r.starts_with('*') || r.starts_with('?') || r.starts_with('!') || r.starts_with('^')
-            {
-                (&r[1..], &r[..1])
-            } else {
-                (r, "")
-            };
-        // Handle $_ special
-        if input.starts_with('$')
-            && r.starts_with('_')
-            && (r.len() == 1
-                || !r
-                    .as_bytes()
-                    .get(1)
-                    .is_some_and(|c| c.is_ascii_alphanumeric()))
-        {
-            return Ok((&r[1..], "_".to_string()));
-        }
-        // Handle $/ special
-        if input.starts_with('$') && r.starts_with('/') {
-            return Ok((&r[1..], "/".to_string()));
-        }
-        // Handle callable operator references: &infix:<...>, &prefix:<...>, ...
-        if input.starts_with('&') {
-            for op_kind in ["infix", "prefix", "postfix", "term", "circumfix"] {
-                if let Some(after_kind) = r.strip_prefix(op_kind)
-                    && let Some(after_colon) = after_kind.strip_prefix(':')
-                {
-                    if let Some(after_open) = after_colon.strip_prefix("<<")
-                        && let Some(end) = after_open.find(">>")
-                    {
-                        let symbol = &after_open[..end];
-                        return Ok((&after_open[end + 2..], format!("{op_kind}:<{symbol}>")));
-                    }
-                    if let Some(after_open) = after_colon.strip_prefix('<')
-                        && let Some(end) = after_open.find('>')
-                    {
-                        let symbol = &after_open[..end];
-                        return Ok((&after_open[end + 1..], format!("{op_kind}:<{symbol}>")));
-                    }
-                }
-            }
-        }
-        // Handle $! as a special variable (not twigil + name)
-        if input.starts_with('$') && twigil == "!" {
-            // If no identifier follows, this is the $! error variable
-            if r.is_empty() || !r.chars().next().is_some_and(is_raku_identifier_start) {
-                return Ok((r, "!".to_string()));
-            }
-        }
-        if twigil.is_empty()
-            && let Some(after_bracket) = r.strip_prefix("::<")
-            && let Some(end) = after_bracket.find('>')
-        {
-            let symbol = &after_bracket[..end];
-            if !symbol.is_empty() {
-                return Ok((&after_bracket[end + 1..], format!("::<{symbol}>")));
-            }
-        }
-        // A fatal qualified-name error (e.g. a null component `$a::::b`) must
-        // propagate, not fall through to the `$foo::` / anonymous-variable arms.
-        if let Err(e) = qualified_ident(r)
-            && e.is_fatal()
-        {
-            return Err(e);
-        }
-        // Handle bare $ (anonymous variable) — no name after sigil
-        if let Ok((mut rest, mut name)) = qualified_ident(r) {
-            while rest.starts_with(':') && !rest.starts_with("::") {
-                let after_colon = &rest[1..];
-                if let Ok((r2, suffix)) = ident(after_colon) {
-                    name.push(':');
-                    name.push_str(&suffix);
-                    rest = r2;
-                    // Double colon after adverb is X::Syntax::Confused
-                    if rest.starts_with("::") {
-                        return Err(PError::fatal(
-                            "X::Syntax::Confused: Confused (double colon in variable adverb)"
-                                .to_string(),
-                        ));
-                    }
-                    // Parse adverb value: <...>, «...», [...], (...)
-                    if let Some((canonical, r3)) = super::primary::var::parse_adverb_value_pub(rest)
-                    {
-                        name.push_str(&canonical);
-                        rest = r3;
-                    }
-                } else {
-                    break;
-                }
-            }
-            let full = if twigil.is_empty() {
-                name
-            } else {
-                format!("{}{}", twigil, name)
-            };
-            Ok((rest, full))
-        } else if let Ok((after_ident, ident_name)) = ident(r)
-            && after_ident.starts_with("::")
-            && twigil.is_empty()
-        {
-            // Variable with trailing :: (e.g., `$foo::`) — package-stash variable.
-            // In Raku, `my $foo::` declares a variable named `$foo::` (distinct from `$foo`).
-            let rest = &after_ident[2..];
-            Ok((rest, format!("{}::", ident_name)))
-        } else if input.starts_with('$') && twigil.is_empty() {
-            // Anonymous scalar variable: bare $
-            Ok((r, "__ANON_STATE__".to_string()))
-        } else if input.starts_with('@') && twigil.is_empty() {
-            Ok((r, "__ANON_ARRAY__".to_string()))
-        } else if input.starts_with('%') && twigil.is_empty() {
-            Ok((r, "__ANON_HASH__".to_string()))
-        } else if input.starts_with('&') && twigil.is_empty() {
-            Ok((r, "__ANON_CODE__".to_string()))
-        } else {
-            Err(PError::expected("variable name"))
-        }
-    } else {
-        Err(PError::expected("variable name"))
-    }
-}
-
-/// Parse a block: { stmts }
-/// Pushes/pops a lexical import scope so that `use` inside a block
-/// only affects that block and its children.
-fn block(input: &str) -> PResult<'_, Vec<Stmt>> {
-    let (input, _) = parse_char(input, '{')?;
-    simple::push_scope();
-    let result = block_inner(input);
-    simple::pop_scope();
-    result
-}
-
-pub(super) fn block_inner(input: &str) -> PResult<'_, Vec<Stmt>> {
-    let (input, stmts) = stmt_list_with_mode(input, false, true)?;
-    let (input, _) = ws(input)?;
-    let (input, _) = parse_char(input, '}')?;
-    Ok((input, stmts))
-}
-
-/// Public accessor for stmt_list (used by primary.rs for block expressions).
-pub(in crate::parser) fn stmt_list_pub(input: &str) -> PResult<'_, Vec<Stmt>> {
-    stmt_list_with_mode(input, false, false)
-}
-
-/// Public accessor for ident (used by primary.rs for hash literal detection).
-pub(super) fn ident_pub(input: &str) -> PResult<'_, String> {
-    ident(input)
-}
-
-/// Public accessor for `temp_stmt` (used by `container.rs` for `(temp @a)` parsing).
-pub(super) fn temp_stmt_pub(input: &str) -> PResult<'_, crate::ast::Stmt> {
-    simple_expr_stmt::temp_stmt(input)
-}
-
-/// Public accessor for `let_stmt` (used by `container.rs` for `(let $x)` parsing).
-pub(super) fn let_stmt_pub(input: &str) -> PResult<'_, crate::ast::Stmt> {
-    simple_expr_stmt::let_stmt(input)
-}
-
-/// Public accessor for var_name (used by primary.rs for anon sub params).
-#[allow(dead_code)]
-pub(super) fn var_name_pub(input: &str) -> PResult<'_, String> {
-    var_name(input)
-}
-/// Public accessor for parse_param_list (used by primary.rs for arrow lambda sub-signatures).
-pub(super) fn parse_param_list_pub(input: &str) -> PResult<'_, Vec<crate::ast::ParamDef>> {
-    parse_param_list(input)
-}
-
-/// Public accessor for parse_param_list_with_return (used by signature literals).
-pub(super) fn parse_param_list_with_return_pub(
-    input: &str,
-) -> PResult<'_, (Vec<crate::ast::ParamDef>, Option<String>)> {
-    sub::parse_param_list_with_return(input)
-}
-
-pub(super) fn parse_return_type_annotation_pub(input: &str) -> PResult<'_, String> {
-    sub::parse_return_type_annotation(input)
-}
-
-pub(super) fn parse_sub_traits_pub(input: &str) -> PResult<'_, sub::SubTraits> {
-    sub::parse_sub_traits(input)
-}
-
-pub(super) fn parse_sub_name_pub(input: &str) -> PResult<'_, String> {
-    sub::parse_sub_name(input)
-}
-
-/// Public accessor for sub_decl_with_semicolon_mode (used by primary/ident.rs for multi sub exprs).
-pub(super) fn sub_decl_with_semicolon_mode_pub(
-    input: &str,
-    allow_main_semicolon_decl: bool,
-) -> PResult<'_, Stmt> {
-    sub::sub_decl_with_semicolon_mode(input, allow_main_semicolon_decl)
-}
-
-/// Public accessor for constant declaration parser (used by primary.rs in expression context).
-pub(super) fn constant_decl_pub(input: &str) -> PResult<'_, Stmt> {
-    decl::constant_decl(input)
-}
-
-/// Public accessor for statement (used by primary.rs for do-statement expressions).
-pub(super) fn statement_pub(input: &str) -> PResult<'_, Stmt> {
-    statement(input)
-}
-
-/// Parse `my`/`our`/`state` in expression context — does not consume
-/// semicolons or apply statement modifiers.
-pub(super) fn my_decl_expr_pub(input: &str) -> PResult<'_, Stmt> {
-    decl::my_decl_expr(input)
-}
-
-/// Public accessor for `for` statement parser (used by primary.rs for `for` expressions).
-pub(super) fn for_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::for_stmt(input)
-}
-
-/// Public accessor for `lazy for` statement parser (used by primary.rs for `lazy for` expressions).
-/// Input should start at `for` (not `lazy`).
-pub(super) fn lazy_for_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::lazy_for_body(input)
-}
-
-/// Public accessor for `if` statement parser (used by primary.rs for `if` expressions).
-pub(super) fn if_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::if_stmt(input)
-}
-
-/// Public accessor for `while` statement parser (used by primary.rs for `while` expressions).
-pub(super) fn while_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::while_stmt(input)
-}
-
-/// Public accessor for `until` statement parser (used by primary.rs for `until` expressions).
-pub(super) fn until_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::until_stmt(input)
-}
-
-/// Public accessor for `loop` statement parser (used by primary.rs for `loop` expressions).
-pub(super) fn loop_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::loop_stmt(input)
-}
-
-pub(super) fn labeled_loop_stmt_pub(input: &str) -> PResult<'_, Stmt> {
-    control::labeled_loop_stmt(input)
-}
-
-/// Parse statements in best-effort mode: return all successfully parsed
-/// statements even if a parse error is encountered partway through.
-/// When a statement fails to parse, skip forward to the next statement
-/// boundary and continue parsing. Used by `load_module` so that
-/// partially-parseable `.rakumod` files still export their functions.
-pub(super) fn stmt_list_partial(input: &str) -> (Vec<Stmt>, Option<String>) {
-    let mut stmts = Vec::new();
-    let mut rest = input;
-    while let Ok((r, _)) = ws_bol(rest) {
-        let r = consume_semicolons(r);
-        let Ok((r, _)) = ws_bol(r) else { break };
-        if r.is_empty() || r.starts_with('}') {
-            break;
-        }
-        match statement(r) {
-            Ok((r, stmt)) => {
-                stmts.push(stmt);
-                rest = r;
-            }
-            Err(_) => {
-                // Skip to next statement boundary and continue.
-                match skip_to_next_statement(r) {
-                    Some(next) => rest = next,
-                    None => break,
-                }
-            }
-        }
-    }
-    (stmts, None)
-}
-
-/// Advance past the current unparseable statement by tracking brace/paren
-/// depth and stopping after a `;` or `}` at depth-0 that likely marks the
-/// end of the current statement.
-fn skip_to_next_statement(input: &str) -> Option<&str> {
-    let bytes = input.as_bytes();
-    let len = bytes.len();
-    let mut brace_depth: i32 = 0;
-    let mut paren_depth: i32 = 0;
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-    let mut prev_was_backslash = false;
-    let mut i = 0;
-    while i < len {
-        let c = bytes[i] as char;
-        if prev_was_backslash {
-            prev_was_backslash = false;
-            i += 1;
-            continue;
-        }
-        if c == '\\' && (in_single_quote || in_double_quote) {
-            prev_was_backslash = true;
-            i += 1;
-            continue;
-        }
-        // Only treat ' as quote delimiter if it's not part of an identifier
-        // (e.g., "doesn't-hang" should not toggle quoting)
-        if c == '\'' && !in_double_quote {
-            let is_identifier_apostrophe = if i > 0 {
-                let prev = bytes[i - 1];
-                prev.is_ascii_alphanumeric() || prev == b'-' || prev == b'_'
-            } else {
-                false
-            } && if i + 1 < len {
-                let next = bytes[i + 1];
-                next.is_ascii_alphanumeric() || next == b'-' || next == b'_'
-            } else {
-                false
-            };
-            if !is_identifier_apostrophe {
-                in_single_quote = !in_single_quote;
-                i += 1;
-                continue;
-            }
-        }
-        if c == '"' && !in_single_quote {
-            in_double_quote = !in_double_quote;
-            i += 1;
-            continue;
-        }
-        if in_single_quote || in_double_quote {
-            i += 1;
-            continue;
-        }
-        match c {
-            '{' => brace_depth += 1,
-            '}' => {
-                if brace_depth > 0 {
-                    brace_depth -= 1;
-                }
-                if brace_depth == 0 && paren_depth == 0 {
-                    // End of a block — skip past it
-                    let after = &input[i + 1..];
-                    let after = after.trim_start();
-                    // Consume optional trailing semicolons
-                    return Some(consume_semicolons(after));
-                }
-            }
-            '(' => paren_depth += 1,
-            ')' => {
-                if paren_depth > 0 {
-                    paren_depth -= 1;
-                }
-            }
-            ';' if brace_depth == 0 && paren_depth == 0 => {
-                return Some(&input[i + 1..]);
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-    None
-}
-
-/// Check if a statement ends with a block (closing `}`).
-/// These "statement-ending blocks" can omit the semicolon only if followed
-/// by a newline, semicolon, closing brace, comma, or end-of-input.
-fn is_block_ending_stmt(stmt: &Stmt) -> bool {
-    match stmt {
-        // Sub declarations only count if they have a body (not forward declarations).
-        // sub_decl does NOT call parse_statement_modifier, so trailing separators
-        // remain in the rest for us to check.
-        Stmt::SubDecl { body, .. } => !body.is_empty(),
-        // `when`/`default` block statements take a `{...}` body and, unlike
-        // `if`/`for`/etc., are not followed by any same-line continuation
-        // keyword (no `else`/`elsif`) and do not consume trailing statement
-        // modifiers. So `when 1 { } when 2 { }` on one line (no `;`) is a
-        // "Strange text after block" error in Raku, matching `given $x { ... }`.
-        Stmt::When { .. } | Stmt::Default(_) => true,
-        // Note: MethodDecl is excluded because it can arrive via my_decl which calls
-        // parse_statement_modifier, consuming the trailing separator before we see it.
-        // Stmt::Block, Stmt::If, Stmt::For, ClassDecl, RoleDecl, Package, etc. are
-        // also excluded for similar reasons -- they may go through paths that consume
-        // trailing separators before returning to stmt_list_with_mode.
-        _ => false,
-    }
-}
-
-/// Check that the remaining input has a proper statement separator
-/// (newline, semicolon, comma, closing brace, or end-of-input)
-/// before any non-whitespace code on the same line.
-fn has_statement_separator(rest: &str) -> bool {
-    for ch in rest.chars() {
-        match ch {
-            // Horizontal whitespace -- keep scanning
-            ' ' | '\t' => continue,
-            // Valid separators: newline, semicolon, closing brace, comma,
-            // or `#` (start of line comment, effectively end-of-statement)
-            '\n' | '\r' | ';' | '}' | ',' | '#' => return true,
-            // Any other non-whitespace on the same line -> missing separator
-            _ => return false,
-        }
-    }
-    // End of input is a valid separator
-    true
-}
-
-fn stmt_list_with_mode(
-    input: &str,
-    allow_mainline_capture: bool,
-    emit_setline: bool,
-) -> PResult<'_, Vec<Stmt>> {
-    let mut stmts = Vec::new();
-    let mut rest = input;
-    let mut saw_compunit_declarator = false;
-    loop {
-        // Use ws_bol (beginning-of-line) so that Pod blocks at the start of
-        // the file or after newlines are consumed.  The regular ws() defaults
-        // to at_line_start=false, which is correct for mid-line positions.
-        let (r, _) = ws_bol(rest)?;
-        // Consume any standalone semicolons
-        let r = consume_semicolons(r);
-        let (r, _) = ws_bol(r)?;
-        // End of block or input
-        if r.is_empty() || r.starts_with('}') {
-            return Ok((r, stmts));
-        }
-        if allow_mainline_capture
-            && let Ok((after_decl, mut main_sub)) = sub::top_level_main_semicolon_decl(r)
-        {
-            if saw_compunit_declarator {
-                return Err(PError::raw(
-                    "X::UnitScope::TooLate: A unit-scoped sub declaration is too late in the compilation unit"
-                        .to_string(),
-                    Some(r.len()),
-                ));
-            }
-            let (tail_rest, tail_stmts) = stmt_list_with_mode(after_decl, false, emit_setline)?;
-            if let Stmt::SubDecl { body, .. } = &mut main_sub {
-                *body = tail_stmts;
-            }
-            stmts.push(main_sub);
-            return Ok((tail_rest, stmts));
-        }
-        if allow_mainline_capture
-            && let Some(after_multi) = keyword("multi", r)
-            && let Ok((after_multi, _)) = ws1(after_multi)
-            && sub::top_level_main_semicolon_decl(after_multi).is_ok()
-        {
-            return Err(PError::raw(
-                "X::UnitScope::Invalid: A unit-scoped sub definition is not allowed except on a MAIN sub; \
-                 Please use the block form. If you did not mean to declare a unit-scoped sub, \
-                 perhaps you accidentally placed a semicolon after routine's definition?"
-                    .to_string(),
-                Some(r.len()),
-            ));
-        }
-        // `unit class`/`unit role`/`unit grammar` (semicolon form): the rest of
-        // the compilation unit is the body of the declared type — capture it
-        // like `unit sub MAIN`. Without this, a top-level `has`/`method` after
-        // `unit class Foo;` is seen as outside any package and rejected with
-        // X::Attribute::NoPackage. (`unit module`/`unit package` are excluded:
-        // the compiler keeps their package context for the rest of the scope.)
-        if allow_mainline_capture && starts_unit_class_role_grammar(r) {
-            let (after_decl, mut decl) = class::unit_module_stmt(r)?;
-            let (tail_rest, tail_stmts) = stmt_list_with_mode(after_decl, false, emit_setline)?;
-            match &mut decl {
-                Stmt::ClassDecl { body, .. } | Stmt::RoleDecl { body, .. } => *body = tail_stmts,
-                _ => {}
-            }
-            stmts.push(decl);
-            return Ok((tail_rest, stmts));
-        }
-        // Emit source line info for deprecation tracking and error reporting.
-        let line = crate::parser::primary::current_line_number(r);
-        let line_valid = crate::parser::primary::is_within_original_source(r);
-        match statement(r) {
-            Ok((r, stmt)) => {
-                // In Raku, after a statement-ending block (e.g. `sub f { 3 }`),
-                // a semicolon, newline, closing brace, or end-of-input is required
-                // before the next statement.  `sub f { 3 } sub g { 3 }` on a
-                // single line is a syntax error.
-                if is_block_ending_stmt(&stmt) && !has_statement_separator(r) {
-                    return Err(PError::fatal(
-                        "X::Syntax::Confused: Strange text after block (missing semicolon or comma?)"
-                            .to_string(),
-                    ));
-                }
-                if matches!(
-                    stmt,
-                    Stmt::Package { .. } | Stmt::ClassDecl { .. } | Stmt::RoleDecl { .. }
-                ) {
-                    saw_compunit_declarator = true;
-                }
-                if emit_setline && line_valid {
-                    stmts.push(Stmt::SetLine(line));
-                }
-                stmts.push(stmt);
-                rest = consume_trailing_comma(r);
-                // In Raku, a stray `)` before `;` at statement level is allowed
-                // (e.g. `throws-like 'code', Exception, 'msg');`).
-                // Consume `)` followed by `;` as a statement terminator.
-                rest = consume_stray_close_paren(rest);
-            }
-            Err(e) => {
-                if e.is_fatal() {
-                    return Err(e);
-                }
-                let consumed = input.len() - r.len();
-                let line_num = input[..consumed].matches('\n').count() + 1;
-                let context: String = r.chars().take(80).collect();
-                return Err(PError::raw(
-                    format!(
-                        "expected statement at line {} (after {} stmts): {} — near: {:?}",
-                        line_num,
-                        stmts.len(),
-                        e,
-                        context
-                    ),
-                    Some(r.len()),
-                ));
-            }
-        }
-    }
-}
-
-/// Consume zero or more semicolons.
-fn consume_semicolons(mut input: &str) -> &str {
-    // In Raku, a comma can act as a statement separator after block-bearing
-    // declarations (e.g. `multi sub foo() { }, multi sub bar() { }`).
-    while input.starts_with(';') || input.starts_with(',') {
-        input = &input[1..];
-        // Also consume whitespace after separators
-        if let Ok((r, _)) = ws(input) {
-            input = r;
-        }
-    }
-    input
-}
-
-/// Consume an optional trailing comma after a statement, used as a statement
-/// separator after block-bearing declarations in Raku
-/// (e.g. `multi sub foo() { }, multi sub bar() { }`).
-fn consume_trailing_comma(input: &str) -> &str {
-    if let Ok((after_ws, _)) = ws(input)
-        && after_ws.starts_with(',')
-    {
-        let r = &after_ws[1..];
-        if let Ok((after_comma_ws, _)) = ws(r) {
-            return after_comma_ws;
-        }
-        return r;
-    }
-    input
-}
-
-/// Consume a stray `)` followed by `;` at statement level.
-/// In Raku, some roast test patterns use `throws-like 'code', Exception, 'msg');`
-/// where the `)` before `;` is a stray closing paren that should be ignored.
-fn consume_stray_close_paren(input: &str) -> &str {
-    if let Ok((after_ws, _)) = ws(input)
-        && after_ws.starts_with(')')
-    {
-        let after_paren = &after_ws[1..];
-        if let Ok((after_paren_ws, _)) = ws(after_paren)
-            && (after_paren_ws.starts_with(';')
-                || after_paren_ws.is_empty()
-                || after_paren_ws.starts_with('}'))
-        {
-            return after_paren_ws;
-        }
-    }
-    input
-}
-
-pub(super) fn parse_pointy_param_pub(input: &str) -> PResult<'_, crate::ast::ParamDef> {
-    control::parse_pointy_param(input)
 }
 
 /// Statement parser function type.
@@ -912,1487 +230,14 @@ pub(super) fn program(input: &str) -> PResult<'_, Vec<Stmt>> {
     };
     // Enable SetLine emission so that source line tracking works for
     // backtraces and $?LINE at the top level.
-    stmt_list_with_mode(input, true, true)
+    stmtlist::stmt_list_with_mode(input, true, true)
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::ast::{CallArg, Expr};
-    use crate::token_kind::TokenKind;
-    use crate::value::Value;
-
-    #[test]
-    fn parse_use_test() {
-        let (rest, stmts) = program("use Test;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::Use { module, .. } if module == "Test"));
-    }
-
-    #[test]
-    fn parse_block_semicolon_expr() {
-        // `{ 1 }; 2` should parse: block followed by semicolon then expression
-        let result = program("{ 1 }; 2");
-        assert!(
-            result.is_ok(),
-            "Failed to parse '{{ 1 }}; 2': {:?}",
-            result.err()
-        );
-        let (rest, stmts) = result.unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-    }
-
-    #[test]
-    fn parse_block_same_line_no_separator_errors() {
-        // `sub f { 3 } sub g { 3 }` should fail
-        let result = program("sub f { 3 } sub g { 3 }");
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn parse_no_strict() {
-        let (rest, stmts) = program("no strict;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::No { module, .. } if module == "strict"));
-    }
-
-    #[test]
-    fn parse_plan_call() {
-        let (rest, stmts) = program("use Test;\nplan 1;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        assert!(matches!(&stmts[1], Stmt::Call { name, .. } if name == "plan"));
-    }
-
-    #[test]
-    fn parse_ok_call_with_named() {
-        let (rest, stmts) = program("use Test;\nok 0, :todo(1);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        if let Stmt::Call { name, args } = &stmts[1] {
-            assert_eq!(name, "ok");
-            assert!(args.len() >= 2);
-            assert!(
-                args.iter()
-                    .any(|arg| matches!(arg, CallArg::Named { name, .. } if name == "todo"))
-            );
-        } else {
-            panic!("Expected Call");
-        }
-    }
-
-    #[test]
-    fn parse_is_deeply_unicode_ascii_minus_complex_args() {
-        let src = "use Test;\nis-deeply −<42+2i>, -<42+2i>, 'prefix, Complex';";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        if let Stmt::Call { name, args } = &stmts[1] {
-            assert_eq!(name, "is-deeply");
-            assert!(args.len() >= 2);
-            assert!(matches!(
-                args[0],
-                CallArg::Positional(Expr::Unary {
-                    op: crate::token_kind::TokenKind::Minus,
-                    ..
-                })
-            ));
-            assert!(matches!(
-                args[1],
-                CallArg::Positional(Expr::Unary {
-                    op: crate::token_kind::TokenKind::Minus,
-                    ..
-                })
-            ));
-        } else {
-            panic!("Expected Call");
-        }
-    }
-
-    #[test]
-    fn parse_is_deeply_unicode_ascii_minus_complex_args_in_block() {
-        let src = "use Test;\n{ is-deeply −<42+2i>, -<42+2i>, 'prefix, Complex'; }";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        assert!(matches!(stmts[1], Stmt::Block(_)));
-    }
-
-    #[test]
-    fn parse_my_var_decl() {
-        let (rest, stmts) = program("my $x = '0';").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { name, .. } if name == "x"));
-    }
-
-    #[test]
-    fn parse_my_var_decl_with_of_type_constraint() {
-        let (rest, stmts) = program("my R1 of Int $x = 1;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(
-            &stmts[0],
-            Stmt::VarDecl {
-                name,
-                type_constraint: Some(tc),
-                ..
-            } if name == "x" && tc == "R1[Int]"
-        ));
-    }
-
-    #[test]
-    fn parse_my_grouped_decl_with_type_smiley() {
-        let (rest, stmts) = program("my Int:D ($x = 5);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::SyntheticBlock(_)));
-    }
-
-    #[test]
-    fn parse_my_invalid_type_smiley_reports_specific_exception() {
-        let err = program("my Int:foo $a;").unwrap_err();
-        assert!(err.message().contains("Invalid type smiley"));
-    }
-
-    #[test]
-    fn parse_my_subset_decl_as_single_statement() {
-        let (rest, stmts) = program("my subset S-Int of Int;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(
-            matches!(&stmts[0], Stmt::SubsetDecl { name, base, .. } if name.resolve() == "S-Int" && base == "Int")
-        );
-    }
-
-    #[test]
-    fn parse_method_arg_hyphenated_subset_name() {
-        let src = "use Test; { my subset S-Int of Int; my subset S-Str of Str; nok S-Int.isa(S-Str), 'isa subset'; }";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-    }
-
-    #[test]
-    fn parse_pointy_param_where_constraint() {
-        let (rest, stmts) = program("-> $ where Int|Bool { }(one True);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-    }
-
-    #[test]
-    fn parse_plan_skip_all() {
-        let (rest, stmts) = program("use Test;\nplan skip-all => \"msg\";").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        if let Stmt::Call { name, args } = &stmts[1] {
-            assert_eq!(name, "plan");
-            assert!(matches!(&args[0], CallArg::Named { name, .. } if name == "skip-all"));
-        } else {
-            panic!("Expected Call");
-        }
-    }
-
-    #[test]
-    fn parse_basic_test_program() {
-        let input = "use Test;\nplan 1;\nok 1, 'test';";
-        let (rest, stmts) = program(input).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 3);
-    }
-
-    #[test]
-    fn parse_if_else() {
-        let (rest, stmts) = program("if $x { say 1; } else { say 2; }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::If { .. }));
-    }
-
-    #[test]
-    fn parse_if_elsif_with_binding_var() {
-        let (rest, stmts) = program("if 0 -> $a { } elsif 1 -> $b { } else { }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Stmt::If {
-                binding_var,
-                else_branch,
-                ..
-            } => {
-                assert_eq!(binding_var.as_deref(), Some("a"));
-                assert!(matches!(
-                    else_branch.first(),
-                    Some(Stmt::If {
-                        binding_var: Some(next),
-                        ..
-                    }) if next == "b"
-                ));
-            }
-            _ => panic!("Expected If statement"),
-        }
-    }
-
-    #[test]
-    fn parse_if_else_with_binding_var_desugars_to_var_decl() {
-        let (rest, stmts) = program("if 0 { } else -> $c { say $c; }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Stmt::If {
-                binding_var,
-                else_branch,
-                ..
-            } => {
-                let source = binding_var.as_ref().expect("generated if binding var");
-                assert!(source.starts_with("$__mutsu_if_bind_"));
-                assert!(matches!(
-                    else_branch.first(),
-                    Some(Stmt::VarDecl { name, expr, .. })
-                        if name == "c"
-                            && matches!(
-                                expr,
-                                Expr::Var(var) if var == source.trim_start_matches('$')
-                            )
-                ));
-            }
-            _ => panic!("Expected If statement"),
-        }
-    }
-
-    #[test]
-    fn parse_for_loop() {
-        let (rest, stmts) = program("for 1..10 -> $i { say $i; }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_loop_with_rw_param_trait() {
-        let (rest, stmts) = program("for @a -> $x is rw { $x++ }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_loop_with_parenthesized_pointy_params() {
-        let (rest, stmts) = program("for @a, @b -> ($x, $y) { say $x; say $y; }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(
-            &stmts[0],
-            Stmt::For {
-                param: Some(name),
-                param_def,
-                ..
-            } if name == "__for_unpack"
-                && matches!(
-                    param_def.as_ref(),
-                    Some(def) if def.sub_signature.as_ref().is_some_and(|sig| sig.len() == 2)
-                )
-        ));
-    }
-
-    #[test]
-    fn parse_compound_assign_on_method_lhs() {
-        let (rest, stmts) = program("for @a -> { .key //= ++$i }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-    }
-
-    #[test]
-    fn parse_topic_method_assign_with_statement_modifier() {
-        let (rest, stmts) = program(".key = 1 for @a;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-    }
-
-    #[test]
-    fn parse_statement_modifier_for_with_trailing_comma() {
-        let (rest, stmts) = program("my @x = gather { take $_ for 1, 2, }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_user_prefix_sub_with_looser_trait_as_full_rhs_expression() {
-        let src = "sub prefix:<pIO> ($p) is looser(&[~]) { $p.IO }\nmy $pre = 'foo';\nsay pIO  $pre ~ '_BAR';";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 3);
-        match &stmts[2] {
-            Stmt::Say(args) => {
-                assert_eq!(args.len(), 1);
-                assert!(matches!(
-                    &args[0],
-                    Expr::Call { name, args }
-                        if name == "prefix:<pIO>"
-                            && args.len() == 1
-                            && matches!(
-                                &args[0],
-                                Expr::Binary { op, .. } if matches!(op, TokenKind::Tilde)
-                            )
-                ));
-            }
-            _ => panic!("expected say statement"),
-        }
-    }
-
-    #[test]
-    fn parse_block_with_unless_statement_modifier() {
-        let (rest, stmts) = program("{ my $a = 1; } unless False;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(
-            &stmts[0],
-            Stmt::If {
-                then_branch,
-                else_branch,
-                ..
-            } if then_branch.len() == 1
-                && matches!(&then_branch[0], Stmt::Block(_))
-                && else_branch.is_empty()
-        ));
-    }
-
-    #[test]
-    fn parse_parenthesized_assign_lvalue_stmt() {
-        let (rest, stmts) = program("($x = $y) = 5;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Stmt::Block(stmts) => {
-                assert_eq!(stmts.len(), 2);
-                assert!(matches!(
-                    &stmts[0],
-                    Stmt::Expr(Expr::AssignExpr { name, expr, .. })
-                        if name == "x" && matches!(expr.as_ref(), Expr::Var(n) if n == "y")
-                ));
-                assert!(matches!(
-                    &stmts[1],
-                    Stmt::Assign { name, expr, .. }
-                        if name == "x" && matches!(expr, Expr::Literal(Value::Int(5)))
-                ));
-            }
-            other => panic!("expected block stmt, got {:?}", other),
-        }
-    }
-
-    #[test]
-    fn parse_parenthesized_bind_lvalue_stmt() {
-        let (rest, stmts) = program("($rv1, $rv2) := |(t2);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::Block(_)));
-    }
-
-    #[test]
-    fn parse_for_loop_with_optional_pointy_param() {
-        let (rest, stmts) = program("for 1..5 -> $x, $y? { say $x }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_loop_with_default_pointy_param() {
-        let (rest, stmts) = program("for 1..5 -> $x, $y = 7 { say $x }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_loop_with_pointy_return_type() {
-        let (rest, stmts) = program("for ^10 -> $_ --> List { take $_ }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_with_typed_unpacking_param_signature() {
-        let (rest, stmts) = program("for a => 1 -> Pair $p (:$key, :$value) { $p }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_with_anonymous_unpacking_param_signature() {
-        let (rest, stmts) = program("for A.new -> $ (:$x) { $x }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_for_with_parenthesized_positional_destructuring_param() {
-        let (rest, stmts) = program("for @pairs -> ($a, $b) { $a }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Stmt::For {
-                param: Some(param),
-                param_def,
-                ..
-            } => {
-                assert_eq!(param, "__for_unpack");
-                let def = param_def.as_ref().as_ref().expect("for param_def");
-                let sub = def.sub_signature.as_ref().expect("for sub_signature");
-                assert_eq!(sub.len(), 2);
-            }
-            _ => panic!("expected for statement with destructuring param"),
-        }
-    }
-
-    #[test]
-    fn parse_for_with_parenthesized_iterable_expression() {
-        let (rest, stmts) = program("for (@a) { .say }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        match &stmts[0] {
-            Stmt::For { iterable, .. } => {
-                assert!(matches!(iterable, Expr::ArrayVar(name) if name == "a"));
-            }
-            _ => panic!("expected for statement"),
-        }
-    }
-
-    #[test]
-    fn parse_for_with_operator_code_ref_in_iterable_list() {
-        let src = "for &infix:<<(==)>>, \"(==)\", &infix:<≡>, \"≡\" -> &op, $name { }";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_chained_inline_modifiers_in_paren_expr() {
-        let (rest, stmts) = program("my @odd = ($_ * $_ if $_ % 2 for 0..10);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_for_expression_in_parens() {
-        let (rest, stmts) = program("my $x = (for ^2 { 41; 42 });").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_topic_mutating_method_stmt() {
-        let (rest, stmts) = program(".=fmt('%03b');").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        // .=method on topic assigns back to $_
-        assert!(matches!(&stmts[0], Stmt::Assign { name, .. } if name == "_"));
-    }
-
-    #[test]
-    fn parse_paren_expr_mutating_method_stmt() {
-        let (rest, stmts) = program("(class { method foo() { self } }.new).=foo;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        // .= on non-lvalue expression is parsed as a method call in expression context
-        assert!(matches!(&stmts[0], Stmt::Expr(Expr::MethodCall { .. })));
-    }
-
-    #[test]
-    fn parse_my_class_expr_in_parens() {
-        let (rest, stmts) =
-            program("(my class :: does Real { method Num { 42e0 } }.new.Bridge);").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::Expr(_)));
-    }
-
-    #[test]
-    fn parse_gather_for_expression() {
-        let (rest, stmts) = program("my $x = (gather for ^5 { take 1 });").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_labeled_for_expression() {
-        let (rest, stmts) = program("my @x = (MEOW: for ^2 { 1 });").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_sub_param_dollar_question() {
-        let (rest, stmts) = program("sub foo($?) { 1 }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::SubDecl { .. }));
-    }
-
-    #[test]
-    fn parse_topic_amp_call_inside_for_block() {
-        let (rest, stmts) = program("sub foo($?) { 1 }; for 1 { .&foo() };").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        assert!(matches!(&stmts[0], Stmt::SubDecl { .. }));
-        assert!(matches!(&stmts[1], Stmt::For { .. }));
-    }
-
-    #[test]
-    fn parse_block_valued_colonpair() {
-        let (rest, stmts) = program("my $x = :out{ .contains('ok') };").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_block_valued_colonpair_with_topic_amp_call_colon_args() {
-        let (rest, stmts) =
-            program("my $x = :out{ .&output-has: :0noks, :0fails, :0todos };").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-    }
-
-    #[test]
-    fn parse_my_regex_decl() {
-        let (rest, stmts) = program("my regex rx { abc };").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(
-            &stmts[0],
-            Stmt::TokenDecl { name, body, .. }
-                if name == "rx"
-                    && matches!(
-                        body.as_slice(),
-                        [Stmt::Expr(Expr::Literal(Value::Regex(pattern)))] if pattern.as_str() == "abc"
-                    )
-        ));
-    }
-
-    #[test]
-    fn parse_sub_decl() {
-        let (rest, stmts) = program("sub foo($x) { return $x; }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::SubDecl { name, .. } if name == "foo"));
-    }
-
-    #[test]
-    fn parse_sub_decl_with_compact_anonymous_sigil_params() {
-        let (rest, stmts) = program("sub foo($$$$) { 1 }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::SubDecl { param_defs, .. } = &stmts[0] {
-            assert_eq!(param_defs.len(), 4);
-            assert!(
-                param_defs
-                    .iter()
-                    .all(|p| p.name == "__ANON_STATE__" && !p.named && !p.slurpy)
-            );
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_forward_sub_declaration_with_semicolon() {
-        let (rest, stmts) = program("sub foo($$$$); sub foo($$$$) { 1 }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        if let Stmt::SubDecl {
-            body, param_defs, ..
-        } = &stmts[0]
-        {
-            assert!(body.is_empty());
-            assert_eq!(param_defs.len(), 4);
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_declared_postfix_operator_call_in_following_statement() {
-        let (rest, stmts) = program("sub postfix:<!!>($x) { $x }; (4!!, 5!!).join(' ');").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        assert!(matches!(&stmts[0], Stmt::SubDecl { name, .. } if name == "postfix:<!!>"));
-        assert!(matches!(&stmts[1], Stmt::Expr(Expr::MethodCall { .. })));
-    }
-
-    #[test]
-    fn parse_forward_sub_without_signature_is_invalid() {
-        let err = program("sub foo;").expect_err("expected parse error");
-        assert!(err.to_string().contains("X::UnitScope::Invalid"));
-    }
-
-    #[test]
-    fn parse_ternary_statement() {
-        let (rest, stmts) = program("1 ?? 2 !! 3;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::Expr(Expr::Ternary { .. })));
-    }
-
-    #[test]
-    fn parse_main_semicolon_captures_following_mainline() {
-        let (rest, stmts) = program("my @*ARGS = <x>; sub MAIN($a); say $a;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2);
-        assert!(matches!(&stmts[0], Stmt::VarDecl { .. }));
-        if let Stmt::SubDecl { name, body, .. } = &stmts[1] {
-            assert_eq!(name, "MAIN");
-            assert_eq!(body.len(), 1);
-            assert!(matches!(&body[0], Stmt::Say(_)));
-        } else {
-            panic!("expected MAIN SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_unit_main_semicolon_captures_following_mainline() {
-        let (rest, stmts) = program("unit sub MAIN($a); say $a;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::SubDecl { name, body, .. } = &stmts[0] {
-            assert_eq!(name, "MAIN");
-            assert_eq!(body.len(), 1);
-            assert!(matches!(&body[0], Stmt::Say(_)));
-        } else {
-            panic!("expected MAIN SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_main_semicolon_after_module_is_too_late() {
-        let err = program("module AtBeginning {}; sub MAIN;").expect_err("expected parse error");
-        assert!(err.to_string().contains("X::UnitScope::TooLate"));
-    }
-
-    #[test]
-    fn parse_sub_decl_with_typed_slurpy_param() {
-        let (rest, stmts) = program("sub foo (Code *$block) { return $block.(); }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::SubDecl { param_defs, .. } = &stmts[0] {
-            assert_eq!(param_defs.len(), 1);
-            assert!(param_defs[0].slurpy);
-            assert_eq!(param_defs[0].type_constraint.as_deref(), Some("Code"));
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_sub_decl_with_slurpy_code_param() {
-        let (rest, stmts) = program("sub bar (*&block) { return &block.(); }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::SubDecl { param_defs, .. } = &stmts[0] {
-            assert_eq!(param_defs.len(), 1);
-            assert!(param_defs[0].slurpy);
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_multi_sub_with_alternate_signature_chain() {
-        let src = "multi sub infix:<+> (Base $b, Exponent $e) | (Exponent $e, Base $b) { 1 }";
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::SubDecl {
-            multi,
-            signature_alternates,
-            ..
-        } = &stmts[0]
-        {
-            assert!(*multi);
-            assert_eq!(signature_alternates.len(), 1);
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_class_decl() {
-        let (rest, stmts) = program("class Foo { has $.x; method bar { 42 } }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::ClassDecl { name, .. } if name == "Foo"));
-    }
-
-    #[test]
-    fn parse_class_decl_with_is_rw_trait() {
-        let (rest, stmts) = program("class Foo is rw { has $.x }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(
-            &stmts[0],
-            Stmt::ClassDecl {
-                name,
-                class_is_rw: true,
-                ..
-            } if name == "Foo"
-        ));
-    }
-
-    #[test]
-    fn parse_class_decl_with_also_is_rw_trait() {
-        let (rest, stmts) = program("class Foo { has $.x; also is rw; has $.y }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::ClassDecl {
-            name,
-            class_is_rw,
-            body,
-            ..
-        } = &stmts[0]
-        {
-            assert_eq!(name.resolve(), "Foo");
-            assert!(*class_is_rw);
-            assert_eq!(body.len(), 2);
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_class_has_decl_is_readonly_trait() {
-        let (rest, stmts) = program("class Foo { has $.x is readonly }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            assert!(matches!(
-                &body[0],
-                Stmt::HasDecl {
-                    is_readonly: true,
-                    ..
-                }
-            ));
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_class_has_decl_dot_equals_initializer() {
-        let src = r#"class MyHandle {
-            has Buf[uint8] $.data .= new: "x".encode;
-        }"#;
-        let (rest, stmts) = program(src).unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            assert!(matches!(
-                &body[0],
-                Stmt::HasDecl {
-                    default: Some(_),
-                    ..
-                }
-            ));
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_class_attribute_postfix_of_type() {
-        let (rest, stmts) = program("class Foo { has @.a of Int }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            if let Stmt::HasDecl {
-                sigil,
-                type_constraint,
-                ..
-            } = &body[0]
-            {
-                assert_eq!(*sigil, '@');
-                assert_eq!(type_constraint.as_deref(), Some("Int"));
-            } else {
-                panic!("expected HasDecl");
-            }
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_class_attribute_where_constraint() {
-        let (rest, stmts) = program("class Foo { has $.x where * > 0 }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            if let Stmt::HasDecl {
-                where_constraint, ..
-            } = &body[0]
-            {
-                assert!(where_constraint.is_some());
-            } else {
-                panic!("expected HasDecl");
-            }
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_method_decl_with_match_var_param() {
-        let (rest, stmts) = program("class Foo { method TOP($/) { 1 } }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            assert!(matches!(&body[0], Stmt::MethodDecl { name, .. } if name == "TOP"));
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_method_decl_with_typed_invocant_marker() {
-        let (rest, stmts) = program("class A { method AT-KEY(A:D: $key) { 1 } }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            if let Stmt::MethodDecl { param_defs, .. } = &body[0] {
-                assert_eq!(param_defs.len(), 1);
-                assert_eq!(param_defs[0].name, "key");
-            } else {
-                panic!("expected MethodDecl");
-            }
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_method_decl_with_operator_name() {
-        let (rest, stmts) =
-            program("class A { method postcircumfix:<{ }>($key) { %!attrs{$key} } }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            assert!(
-                matches!(&body[0], Stmt::MethodDecl { name, .. } if name == "postcircumfix:<{ }>")
-            );
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_sub_decl_with_typed_invocant_marker() {
-        let (rest, stmts) = program("sub f(A:D: $key) { $key }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::SubDecl { param_defs, .. } = &stmts[0] {
-            assert_eq!(param_defs.len(), 1);
-            assert_eq!(param_defs[0].name, "key");
-        } else {
-            panic!("expected SubDecl");
-        }
-    }
-
-    #[test]
-    fn parse_method_decl_with_explicit_invocant_param() {
-        let (rest, stmts) = program("class Foo { method bar ($self: $num) { $num } }").unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            if let Stmt::MethodDecl { param_defs, .. } = &body[0] {
-                assert_eq!(param_defs.len(), 2);
-                assert_eq!(param_defs[0].name, "self");
-                assert!(param_defs[0].traits.iter().any(|t| t == "invocant"));
-                assert_eq!(param_defs[1].name, "num");
-                assert!(!param_defs[1].traits.iter().any(|t| t == "invocant"));
-            } else {
-                panic!("expected MethodDecl");
-            }
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_method_decl_with_explicit_invocant_and_return_type() {
-        let (rest, stmts) =
-            program("class Foo { method bar ($self: $num --> Foo) returns Foo { $self } }")
-                .unwrap();
-        assert_eq!(rest, "");
-        if let Stmt::ClassDecl { body, .. } = &stmts[0] {
-            if let Stmt::MethodDecl { param_defs, .. } = &body[0] {
-                assert_eq!(param_defs.len(), 2);
-                assert!(param_defs[0].traits.iter().any(|t| t == "invocant"));
-                assert_eq!(param_defs[1].name, "num");
-            } else {
-                panic!("expected MethodDecl");
-            }
-        } else {
-            panic!("expected ClassDecl");
-        }
-    }
-
-    #[test]
-    fn parse_role_decl_with_generics_and_does_clause() {
-        let (rest, stmts) = program("role R2[Cool ::T] does R1[T] is ok { }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::RoleDecl { name, .. } if name == "R2"));
-    }
-
-    #[test]
-    fn parse_grammar_decl_with_does_clause() {
-        let (rest, stmts) = program("grammar G does R { rule TOP { ^ <x> } }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::ClassDecl { name, .. } if name == "G"));
-    }
-
-    #[test]
-    fn parse_token_decl_with_regex_like_body() {
-        let (rest, stmts) = program("token TOP { <fred>+ }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::TokenDecl { name, .. } if name == "TOP"));
-    }
-
-    #[test]
-    fn parse_token_decl_with_sym_variant_name() {
-        let (rest, stmts) = program("token fred:sym<foo> { <sym> \\d+ }").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-        assert!(matches!(&stmts[0], Stmt::TokenDecl { name, .. } if name == "fred:sym<foo>"));
-    }
-
-    #[test]
-    fn parse_make_with_token_term_in_array() {
-        let (rest, stmts) = program("make [token { \"bar\" }]").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-    }
-
-    #[test]
-    fn parse_chained_constant_declarator_rhs() {
-        let (rest, stmts) = program("my $a = constant $b = 42;").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 1);
-    }
-
-    #[test]
-    fn parse_my_no_space() {
-        let (rest, stmts) = program("my $a=0; say $a").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(stmts.len(), 2, "stmts: {:?}", stmts);
-        if let Stmt::VarDecl { name, expr, .. } = &stmts[0] {
-            assert_eq!(name, "a");
-            assert!(
-                matches!(expr, Expr::Literal(Value::Int(0))),
-                "Expected Int(0), got {:?}",
-                expr
-            );
-        } else {
-            panic!("Expected VarDecl, got {:?}", stmts[0]);
-        }
-    }
-
-    #[test]
-    fn statement_memo_hits_on_reparse() {
-        reset_statement_memo();
-        let input = "say 42";
-        let _ = statement_pub(input).unwrap();
-        let (hits1, misses1, stores1) = statement_memo_stats();
-        assert_eq!(hits1, 0);
-        assert!(misses1 >= 1);
-        assert!(stores1 >= 1);
-
-        let _ = statement_pub(input).unwrap();
-        let (hits2, _misses2, stores2) = statement_memo_stats();
-        assert!(hits2 >= 1);
-        assert!(stores2 >= stores1);
-    }
-
-    #[test]
-    fn merge_expected_messages_deduplicates() {
-        let merged = super::super::parse_result::merge_expected_messages(
-            "expected foo",
-            &["foo".to_string(), "bar".to_string()],
-        );
-        assert_eq!(merged, vec!["foo", "bar"]);
-    }
-
-    #[test]
-    fn merge_expected_messages_strips_prefix_consistently() {
-        let merged = super::super::parse_result::merge_expected_messages(
-            "expected alpha",
-            &["beta".to_string(), "gamma".to_string()],
-        );
-        assert_eq!(merged, vec!["alpha", "beta", "gamma"]);
-    }
-
-    #[test]
-    fn statement_modifier_reports_missing_condition() {
-        let base = Stmt::Expr(Expr::Literal(Value::Int(1)));
-        let err = parse_statement_modifier(" if ", base).unwrap_err();
-        assert!(err.message().contains("after 'if'"));
-    }
-
-    #[test]
-    fn assign_stmt_reports_missing_rhs_for_compound_assign() {
-        let err = assign::assign_stmt("$x +=").unwrap_err();
-        assert!(err.message().contains("compound assignment"));
-    }
-
-    #[test]
-    fn assign_stmt_reports_missing_method_name_for_mutating_call() {
-        let err = assign::assign_stmt("$x .=").unwrap_err();
-        assert!(err.message().contains("method name after '.='"));
-    }
-
-    #[test]
-    fn assign_stmt_parses_zip_compound_assign() {
-        let (rest, stmt) = assign::assign_stmt("$a Z+= 2;").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(stmt, Stmt::Assign { .. }));
-    }
-
-    #[test]
-    fn assign_stmt_parses_comma_compound_assign() {
-        let (rest, stmt) = assign::assign_stmt("%h ,= 1;").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(stmt, Stmt::Assign { .. }));
-    }
-
-    #[test]
-    fn assign_stmt_scalar_item_assignment_stops_at_comma() {
-        // `$x = 1, 2` is `($x = 1), 2`: only the first element is assigned, and
-        // the rest become a sink-context comma list.
-        let (rest, stmt) = assign::assign_stmt("$x = 1, 2;").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Expr(Expr::ArrayLiteral(items)) => {
-                assert_eq!(items.len(), 2);
-                assert!(matches!(
-                    &items[0],
-                    Expr::AssignExpr { name, expr, .. }
-                        if name == "x" && matches!(expr.as_ref(), Expr::Literal(Value::Int(1)))
-                ));
-                assert!(matches!(&items[1], Expr::Literal(Value::Int(2))));
-            }
-            other => panic!("expected comma-list expr statement, got {:?}", other),
-        }
-    }
-
-    #[test]
-    fn assign_stmt_scalar_plain_assignment_has_no_comma() {
-        // Without a trailing comma the assignment is a plain `Stmt::Assign`.
-        let (rest, stmt) = assign::assign_stmt("$x = 1;").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            stmt,
-            Stmt::Assign { name, expr, .. }
-                if name == "x" && matches!(expr, Expr::Literal(Value::Int(1)))
-        ));
-    }
-
-    #[test]
-    fn assign_stmt_array_list_assignment_absorbs_comma() {
-        // `@a = 1, 2` keeps the comma-absorbing RHS (list assignment).
-        let (rest, stmt) = assign::assign_stmt("@a = 1, 2;").unwrap();
-        assert_eq!(rest, "");
-        assert!(matches!(
-            stmt,
-            Stmt::Assign { name, expr, .. }
-                if name == "@a" && matches!(&expr, Expr::ArrayLiteral(items) if items.len() == 2)
-        ));
-    }
-
-    #[test]
-    fn assign_expr_parses_reverse_bracket_metaop_assign() {
-        let (rest, expr) = assign::try_parse_assign_expr("$y [R/]= 1").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::AssignExpr { name, expr, .. } => {
-                assert_eq!(name, "y");
-                match *expr {
-                    Expr::MetaOp {
-                        meta,
-                        op,
-                        left,
-                        right,
-                    } => {
-                        assert_eq!(meta, "R");
-                        assert_eq!(op, "/");
-                        assert!(matches!(*left, Expr::Var(ref n) if n == "y"));
-                        assert!(matches!(*right, Expr::Literal(Value::Int(1))));
-                    }
-                    other => panic!("expected meta-op assignment expr, got {other:?}"),
-                }
-            }
-            other => panic!("expected AssignExpr, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn assign_expr_bracket_metaop_leaves_following_call_args() {
-        let (rest, _) = assign::try_parse_assign_expr("$y [R/]= 1, 1/5, 'desc'").unwrap();
-        assert_eq!(rest, ", 1/5, 'desc'");
-    }
-
-    #[test]
-    fn assign_stmt_parses_nested_bracket_metaop_assign() {
-        let (rest, stmt) = assign::assign_stmt("$x [R[+]]= 1;").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Assign { expr, .. } => match expr {
-                Expr::MetaOp { meta, op, .. } => {
-                    assert_eq!(meta, "R");
-                    assert_eq!(op, "+");
-                }
-                other => panic!("expected meta-op assignment stmt, got {other:?}"),
-            },
-            other => panic!("expected Assign stmt, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn assign_expr_parses_unbracketed_cross_metaop_assign() {
-        let (rest, expr) = assign::try_parse_assign_expr("@a X*= 10").unwrap();
-        assert_eq!(rest, "");
-        match expr {
-            Expr::AssignExpr { name, expr, .. } => {
-                assert_eq!(name, "@a");
-                match *expr {
-                    Expr::MetaOp {
-                        meta,
-                        op,
-                        left,
-                        right,
-                    } => {
-                        assert_eq!(meta, "X");
-                        assert_eq!(op, "*");
-                        assert!(matches!(*left, Expr::ArrayVar(ref n) if n == "a"));
-                        assert!(matches!(*right, Expr::Literal(Value::Int(10))));
-                    }
-                    other => panic!("expected meta-op assignment expr, got {other:?}"),
-                }
-            }
-            other => panic!("expected AssignExpr, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn assign_stmt_parses_unbracketed_cross_metaop_assign() {
-        let (rest, stmt) = assign::assign_stmt("@a X*= 10;").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Assign { name, expr, .. } => {
-                assert_eq!(name, "@a");
-                match expr {
-                    Expr::MetaOp { meta, op, .. } => {
-                        assert_eq!(meta, "X");
-                        assert_eq!(op, "*");
-                    }
-                    other => panic!("expected meta-op assignment stmt, got {other:?}"),
-                }
-            }
-            other => panic!("expected Assign stmt, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn assign_stmt_parses_unbracketed_cross_plain_assign() {
-        let (rest, stmt) = assign::assign_stmt("@a X= 10;").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Assign { name, expr, .. } => {
-                assert_eq!(name, "@a");
-                match expr {
-                    Expr::MetaOp { meta, op, .. } => {
-                        assert_eq!(meta, "X");
-                        assert_eq!(op, "=");
-                    }
-                    other => panic!("expected meta-op assignment stmt, got {other:?}"),
-                }
-            }
-            other => panic!("expected Assign stmt, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn parse_meta_compound_assign_op_recognizes_reverse_add_assign() {
-        let parsed = assign::parse_meta_compound_assign_op("R+= 10");
-        assert!(parsed.is_some());
-        let (rest, meta, op) = parsed.unwrap();
-        assert_eq!(rest, " 10");
-        assert_eq!(meta, "R");
-        assert_eq!(op, "+");
-    }
-
-    #[test]
-    fn known_call_stmt_accepts_bracket_metaop_assign_argument() {
-        simple::register_module_exports("Test");
-        let parsed = simple::known_call_stmt("is $y [R/]= 1, 1/5, \"[R/]= works correctly (1)\";");
-        assert!(parsed.is_ok());
-    }
-
-    #[test]
-    fn statement_accepts_known_call_with_bracket_metaop_assign_argument() {
-        simple::register_module_exports("Test");
-        let parsed = statement("is $y [R/]= 1, 1/5, \"[R/]= works correctly (1)\";");
-        assert!(parsed.is_ok());
-    }
-
-    #[test]
-    fn expr_stmt_parses_reverse_assignment_into_var_decl() {
-        let (rest, stmt) = simple::expr_stmt("1 R= my $x").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::VarDecl { name, expr, .. } => {
-                assert_eq!(name, "x");
-                assert!(matches!(expr, Expr::Literal(Value::Int(1))));
-            }
-            other => panic!("expected VarDecl, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn expr_stmt_rejects_reverse_bind_assignment() {
-        let err = simple::expr_stmt("5 R:= $x").unwrap_err();
-        assert!(err.message().contains("Cannot reverse the args of :="));
-    }
-
-    #[test]
-    fn expr_stmt_non_lvalue_assignment_is_not_silently_ignored() {
-        let (rest, stmt) = simple::expr_stmt("(\"a\" R~ \"b\") = 1").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Expr(Expr::Call { name, .. }) => {
-                assert_eq!(name, "__mutsu_assign_callable_lvalue");
-            }
-            other => panic!("expected assignment call expression, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn known_call_stmt_reports_argument_parse_context() {
-        simple::register_module_exports("Test");
-        let err = simple::known_call_stmt("ok ,").unwrap_err();
-        assert!(err.message().contains("known call arguments"));
-    }
-
-    #[test]
-    fn known_call_stmt_reports_missing_comma_argument() {
-        simple::register_module_exports("Test");
-        let err = simple::known_call_stmt("ok(,)").unwrap_err();
-        assert!(err.message().contains("known call arguments"));
-    }
-
-    #[test]
-    fn known_call_stmt_reports_missing_named_argument_value() {
-        simple::register_module_exports("Test");
-        let err = simple::known_call_stmt("ok :foo()").unwrap_err();
-        assert!(err.message().contains("named argument value"));
-    }
-
-    #[test]
-    fn known_call_stmt_parses_unicode_and_ascii_minus_angle_complex_args() {
-        simple::reset_user_subs();
-        simple::register_module_exports("Test");
-        let (rest, stmt) =
-            simple::known_call_stmt("is-deeply −<42+2i>, -<42+2i>, 'prefix, Complex'").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Call { name, args } => {
-                assert_eq!(name, "is-deeply");
-                assert!(args.len() >= 2);
-                assert!(matches!(
-                    args[0],
-                    CallArg::Positional(Expr::Unary {
-                        op: crate::token_kind::TokenKind::Minus,
-                        ..
-                    })
-                ));
-                assert!(matches!(
-                    args[1],
-                    CallArg::Positional(Expr::Unary {
-                        op: crate::token_kind::TokenKind::Minus,
-                        ..
-                    })
-                ));
-            }
-            other => panic!("Expected Call, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn known_call_stmt_keeps_chained_assignment_as_single_argument() {
-        simple::register_module_exports("Test");
-        let (rest, stmt) = simple::known_call_stmt("is $x = $y = $z, 18;").unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Call { name, args } => {
-                assert_eq!(name, "is");
-                assert!(matches!(
-                    args.first(),
-                    Some(CallArg::Positional(Expr::AssignExpr { .. }))
-                ));
-                assert!(matches!(
-                    args.get(1),
-                    Some(CallArg::Positional(Expr::Literal(Value::Int(18))))
-                ));
-            }
-            other => panic!("expected known call statement, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn known_call_stmt_splits_assignment_from_following_test_args() {
-        simple::register_module_exports("Test");
-        let (rest, stmt) = simple::known_call_stmt(
-            "is $ = bar(), \"NoModule::bar\", \"can import symbol not inside module\";",
-        )
-        .unwrap();
-        assert_eq!(rest, "");
-        match stmt {
-            Stmt::Call { name, args } => {
-                assert_eq!(name, "is");
-                assert!(matches!(
-                    args.first(),
-                    Some(CallArg::Positional(Expr::AssignExpr { .. }))
-                ));
-                assert!(matches!(
-                    args.get(1),
-                    Some(CallArg::Positional(Expr::Literal(Value::Str(expected))))
-                        if **expected == "NoModule::bar"
-                ));
-                assert!(matches!(
-                    args.get(2),
-                    Some(CallArg::Positional(Expr::Literal(Value::Str(expected))))
-                        if **expected == "can import symbol not inside module"
-                ));
-            }
-            other => panic!("expected known call statement, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn qualified_ident_requires_segment_after_double_colon() {
-        let err = qualified_ident("Foo::").unwrap_err();
-        assert!(err.message().contains("identifier after '::'"));
-    }
-
-    #[test]
-    fn import_scope_is_lexical() {
-        simple::reset_user_subs();
-        // At top scope, "ok" is not imported
-        assert!(!simple::is_imported_function("ok"));
-        // Push a child scope and import Test
-        simple::push_scope();
-        simple::register_module_exports("Test");
-        assert!(simple::is_imported_function("ok"));
-        // Pop the child scope — "ok" should no longer be visible
-        simple::pop_scope();
-        assert!(
-            !simple::is_imported_function("ok"),
-            "imported function should not be visible after scope pop"
-        );
-    }
-
-    #[test]
-    fn import_scope_inherits_from_parent() {
-        simple::reset_user_subs();
-        // Import at top scope
-        simple::register_module_exports("Test");
-        assert!(simple::is_imported_function("ok"));
-        // Push a child scope — parent imports should still be visible
-        simple::push_scope();
-        assert!(
-            simple::is_imported_function("ok"),
-            "parent scope imports should be visible in child"
-        );
-        simple::pop_scope();
-        // Still visible in parent
-        assert!(simple::is_imported_function("ok"));
-        simple::reset_user_subs();
-    }
-
-    #[test]
-    fn user_sub_scope_is_lexical() {
-        simple::reset_user_subs();
-        assert!(!simple::is_user_declared_sub("my-func"));
-        // Declare a sub in a child scope
-        simple::push_scope();
-        simple::register_user_sub("my-func");
-        assert!(simple::is_user_declared_sub("my-func"));
-        // Pop — sub should no longer be visible
-        simple::pop_scope();
-        assert!(
-            !simple::is_user_declared_sub("my-func"),
-            "user sub should not be visible after scope pop"
-        );
-    }
-
-    #[test]
-    fn user_sub_scope_inherits_from_parent() {
-        simple::reset_user_subs();
-        simple::register_user_sub("outer-func");
-        assert!(simple::is_user_declared_sub("outer-func"));
-        // Child scope should see parent's subs
-        simple::push_scope();
-        assert!(
-            simple::is_user_declared_sub("outer-func"),
-            "parent sub should be visible in child scope"
-        );
-        simple::pop_scope();
-        assert!(simple::is_user_declared_sub("outer-func"));
-        simple::reset_user_subs();
-    }
-
-    #[test]
-    fn done_parses_as_react_done_stmt() {
-        let (rest, stmt) = simple::known_call_stmt("done").unwrap();
-        assert!(rest.is_empty(), "rest should be empty, got: {rest:?}");
-        assert!(
-            matches!(stmt, Stmt::ReactDone),
-            "expected ReactDone, got: {stmt:?}"
-        );
-    }
-
-    #[test]
-    fn done_with_semicolon_parses_as_react_done_stmt() {
-        let (rest, stmt) = simple::known_call_stmt("done;").unwrap();
-        assert!(rest.is_empty(), "rest should be empty, got: {rest:?}");
-        assert!(
-            matches!(stmt, Stmt::ReactDone),
-            "expected ReactDone, got: {stmt:?}"
-        );
-    }
-
-    #[test]
-    fn proceed_parses_as_proceed_stmt() {
-        let (rest, stmt) = simple::known_call_stmt("proceed").unwrap();
-        assert!(rest.is_empty());
-        assert!(matches!(stmt, Stmt::Proceed));
-    }
-
-    #[test]
-    fn succeed_parses_as_succeed_stmt() {
-        let (rest, stmt) = simple::known_call_stmt("succeed").unwrap();
-        assert!(rest.is_empty());
-        assert!(matches!(stmt, Stmt::Succeed));
-    }
-
-    #[test]
-    fn var_name_accepts_root_stash_symbol_declarator() {
-        let (rest, name) = var_name("$::<!@#$>").unwrap();
-        assert_eq!(rest, "");
-        assert_eq!(name, "::<!@#$>");
-    }
-}
+mod tests;
+#[cfg(test)]
+mod tests_2;
+#[cfg(test)]
+mod tests_3;
+#[cfg(test)]
+mod tests_4;
